Menna

+++ Repo/Repository.cs
@@ -1,4 +1,5 @@
 ﻿using invoice.Core.DTO;
+using invoice.Core.Interfaces;
 using invoice.Repo;
 using invoice.Repo.Data;
 using Microsoft.EntityFrameworkCore;
@@ -25,9 +26,13 @@ namespace Repo
 
             if (includes != null && includes.Any())
                 query = includes.Aggregate(query, (current, include) => current.Include(include));
+           
+            if (typeof(T).GetProperty("IsDeleted") != null)
+                query = query.Where(e => EF.Property<bool>(e, "IsDeleted") == false);
 
             if (!string.IsNullOrEmpty(userId) && typeof(T).GetProperty("UserId") != null)
                 query = query.Where(e => EF.Property<string>(e, "UserId") == userId);
+               
 
             return await query.ToListAsync();
         }
@@ -40,10 +45,26 @@ namespace Repo
                 query = include(query);
 
             if (!string.IsNullOrEmpty(userId) && typeof(T).GetProperty("UserId") != null)
-                query = query.Where(e => EF.Property<string>(e, "UserId") == userId);
+                query = query.Where(e => EF.Property<string>(e, "UserId") == userId &&
+                 EF.Property<bool>(e, "IsDeleted") == false);
 
             return await query.FirstOrDefaultAsync(e => EF.Property<string>(e, "Id") == id);
         }
+
+        public async Task<List<T>> GetByIdsAsync(List<string> ids, string userId = null, Func<IQueryable<T>, IQueryable<T>> include = null)
+        {
+            IQueryable<T> query = _dbSet;
+
+            if (include != null)
+                query = include(query);
+
+            if (!string.IsNullOrEmpty(userId) && typeof(T).GetProperty("UserId") != null)
+                query = query.Where(e => EF.Property<string>(e, "UserId") == userId &&
+                                         EF.Property<bool>(e, "IsDeleted") == false);
+
+            return await query.Where(e => ids.Contains(EF.Property<string>(e, "Id"))).ToListAsync();
+        }
+
 
         public async Task<T> GetByUserIdAsync(string userId, Func<IQueryable<T>, IQueryable<T>> include = null)
         {
@@ -52,7 +73,8 @@ namespace Repo
             if (include != null)
                 query = include(query);
 
-            return await query.FirstOrDefaultAsync(e => EF.Property<string>(e, "UserId") == userId);
+            return await query.FirstOrDefaultAsync(e => EF.Property<string>(e, "UserId") == userId &&
+             EF.Property<bool>(e, "IsDeleted") == false);
         }
 
         public async Task<IEnumerable<T>> QueryAsync(Expression<Func<T, bool>> predicate, params Expression<Func<T, object>>[] includes)
@@ -117,7 +139,7 @@ namespace Repo
                 Data = entity
             };
         }
-
+        
         public async Task<GeneralResponse<IEnumerable<T>>> UpdateRangeAsync(IEnumerable<T> entities)
         {
             _dbSet.UpdateRange(entities);
@@ -143,8 +165,18 @@ namespace Repo
                 };
             }
 
-            _dbSet.Remove(entity);
-            await _context.SaveChangesAsync();
+            var property = entity.GetType().GetProperty("IsDeleted");
+            if (property != null)
+            {
+                property.SetValue(entity, true);
+                _dbSet.Update(entity);
+                await _context.SaveChangesAsync();
+            }
+            else
+            {
+                _dbSet.Remove(entity);
+                await _context.SaveChangesAsync();
+            }
 
             return new GeneralResponse<T>
             {
@@ -153,6 +185,7 @@ namespace Repo
                 Data = entity
             };
         }
+      
 
         public async Task<GeneralResponse<IEnumerable<T>>> DeleteRangeAsync(IEnumerable<string> ids)
         {
@@ -167,8 +200,20 @@ namespace Repo
                 };
             }
 
-            _dbSet.RemoveRange(entities);
-            await _context.SaveChangesAsync();
+
+            var property = entities.GetType().GetProperty("IsDeleted");
+            if (property != null)
+            {
+                property.SetValue(entities, true);
+                _dbSet.UpdateRange(entities);
+                await _context.SaveChangesAsync();
+            }
+            else
+            {
+                _dbSet.RemoveRange(entities);
+                await _context.SaveChangesAsync();
+            }
+          
 
             return new GeneralResponse<IEnumerable<T>>
             {
